+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The `verb-json-get` function now accepts negative integer arguments for accessing list elements.
 - Changed default binding of `verb-send-request-on-point-no-window` to <kbd>C-c C-r C-<return></kbd>.
 - Allow using single- or multi-line lambda expressions for `Verb-Map-Request`.
+- The `verb-auto-kill-response-buffers` customizable variable can now be set to an integer. This will cause all response buffers to be killed when a request is sent, except the N most recent ones.
 
 ## **2.15.0** - 2021-11-03 (MELPA Stable)
 - Fixed font locking on indented Babel source blocks.